requirements.txt
absl-py==2.1.0
accelerate==0.29.3
appdirs==1.4.4
astunparse==1.6.3
blobfile==3.0.0
build==1.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.15.4
flatbuffers==24.3.25
flit_core @ file:///tmp/Python/3.11.5/GCCcore-13.2.0/flit_core/flit_core-3.9.0
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
keras-nightly==3.5.0.dev2024091303
kiwisolver==1.4.7
kornia==0.7.1
libclang==18.1.1
lightning-utilities==0.11.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
ml-dtypes==0.4.0
mpi4py==4.0.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.3
opt-einsum==3.3.0
optree==0.12.1
packaging @ file:///tmp/Python/3.11.5/GCCcore-13.2.0/packaging/packaging-23.2
pillow==10.2.0
protobuf==4.25.4
psutil==6.0.0
pycryptodomex==3.20.0
Pygments==2.18.0
pyparsing==3.1.4
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.14.1
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools-scm @ file:///tmp/Python/3.11.5/GCCcore-13.2.0/setuptoolsscm/setuptools-scm-8.0.4
six==1.16.0
smmap==5.0.1
sympy==1.12.1
tb-nightly==2.18.0a20240913
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tomli @ file:///tmp/Python/3.11.5/GCCcore-13.2.0/tomli/tomli-2.0.1
torch==2.4.0+cu124.post2
torchmetrics==1.3.2
torchvision==0.19.0+cu124
tqdm==4.66.2
triton==2.3.1
typing_extensions @ file:///tmp/Python/3.11.5/GCCcore-13.2.0/typing_extensions/typing_extensions-4.8.0
urllib3==2.2.2
wandb==0.16.5
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
z3-solver @ file:///tmp/Z3/4.13.0/GCCcore-13.2.0/z3solver/z3-solver-4.13.0.0
